% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_all}
\alias{data_all}
\title{An example of 2-group mixed data}
\format{
## `data_all`
A data frame with 500 rows and 100 columns.
}
\usage{
data_all
}
\description{
A dataset containing 50 numeric and 50 categorical variables
Includes 250 observations in each group
}
\keyword{datasets}
