% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_mini}
\alias{data_mini}
\title{A toy example of 2-group mixed data}
\format{
## `data_mini`
A data frame with 500 rows and 10 columns.
}
\usage{
data_mini
}
\description{
A dataset containing 4 numeric and 6 categorical variables
Includes 250 observations in each group
}
\keyword{datasets}
