% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ind_disc_mini}
\alias{ind_disc_mini}
\title{A toy example of 2-group mixed data}
\format{
## `ind_disc_mini`
A 6-length vector with discrete variable indices.
}
\usage{
ind_disc_mini
}
\description{
A vector indicating which columns in 'data_mini' have categorical variables
}
\keyword{datasets}
