% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ind_disc}
\alias{ind_disc}
\title{An example of 2-group mixed data}
\format{
## `ind_disc`
A 250-length vector with discrete variable indices.
}
\usage{
ind_disc
}
\description{
An indicator vector which variables are categorical
}
\keyword{datasets}
